tools/picturas-invert-tool-ms/pyproject.toml
[tool.poetry]
name = "picturas-invert-tool-ms"
version = "0.1.0"
description = "PictuRAS microservice (tool) that inverts the colors of an image."
authors = ["José Correia <pg55967@alunos.uminho.pt>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9.15"
pika = "^1.3.2"
pydantic = "^2.10.3"
pillow = "^10.4.0"

[tool.poetry.group.dev.dependencies]
yapf = "^0.43.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
